requirements.txt
absl-py==1.1.0
blobfile==2.0.1
colorama==0.4.3
einops==0.4.1
flax==0.6.6
imageio==2.19.3
lmdb==1.4.0
matplotlib==3.5.2
nbconvert>=6.5.1
numba==0.55.2
numpy==1.22.4
protobuf==3.20.3
resize-right
scikit-image
scipy==1.8.1
setuptools>=65.5.1
tabulate==0.8.9
tensorboard==2.12.2
tensorflow==2.12.0
torch==1.13.1+cu116
torchvision==0.14.1+cu116
tqdm==4.64.0
